Move handleSelectWord logic out of Word.tsx
